rename service folder -> services
